Extract color constants
